+++ stream/stream_model.py
@@ -28,7 +28,7 @@ class StreamModel:
         self.ms_container_commands: str = ''
 
         self.mp_ffmpeg_reader_owner_pid: int = 0
-        self.ffmpeg_reader_frame_rate: int = 1
+        self.ffmpeg_reader_frame_rate: float = 1.0
         self.ffmpeg_reader_width: int = 640
         self.ffmpeg_reader_height: int = 360
 
@@ -41,7 +41,7 @@ class StreamModel:
         # FFmpeg snapshot for AI.
         self.snapshot_enabled: bool = False
         self.snapshot_pid: int = 0
-        self.snapshot_frame_rate: int = 1
+        self.snapshot_frame_rate: float = 1.0
         self.snapshot_width: int = 640
         self.snapshot_height: int = 360
 
@@ -73,7 +73,7 @@ class StreamModel:
         self.ffmpeg_reader_height = source.ffmpeg_reader_height
 
         self.snapshot_enabled = source.snapshot_enabled
-        self.snapshot_frame_rate: int = source.snapshot_frame_rate
+        self.snapshot_frame_rate: float = source.snapshot_frame_rate
         self.snapshot_width: int = source.snapshot_width
         self.snapshot_height: int = source.snapshot_height
